Finished Happiness and Health Meter.
Meters and Play work as intended. Health increases when player clicks one of the buttons from "Play". Decreases when happiness reaches a certain amount over a period of time; if waste meter reaches 10. Happiness increases if Fed or played with; decreases with a percentage of the Hunger meter/time last played.